apps/api/pyproject.toml
[project]
name = "joby-api"
version = "0.1.0"
description = "Joby FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "apscheduler>=3.10",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]
scrapers = [
  # JobSpy aggregates LinkedIn/Indeed/Glassdoor/ZipRecruiter/Google Jobs.
  # Kept optional because it pulls in pandas and other heavy deps; the
  # daemon lazy-imports it so the rest of Joby runs without it installed.
  "python-jobspy>=1.1.80",
]

[project.scripts]
joby = "app.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
